requirements.txt
appdirs==1.4.4
asgiref==3.4.1
bcrypt==3.2.0
black==21.6b0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.1
cryptography==3.4.7
et-xmlfile==1.1.0
fastapi==0.66.0
greenlet==1.1.0
h11==0.12.0
idna==3.3
mypy-extensions==0.4.3
openpyxl==3.0.7
pathspec==0.8.1
pycparser==2.20
pydantic==1.8.2
PyJWT==1.7.1
PyMySQL==1.0.2
python-decouple==3.3
regex==2021.7.6
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.20
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.7
uvicorn==0.14.0
XlsxWriter==3.0.1

